+++ tests/unit-tests/jobs/models/jobManager-test.js
@@ -15,8 +15,6 @@ const databaseConnector = require('../../../../src/jobs/models/database/database
     webhooksManager = require('../../../../src/webhooks/models/webhookManager'),
     basicTest = require('../../../testExamples/Basic_test.json'),
     reportsManager = require('../../../../src/reports/models/reportsManager'),
-    chaosExperimentsManager = require('../../../../src/chaos-experiments/models/chaosExperimentsManager'),
-    chaosExperimentConnector = require('../../../../src/chaos-experiments/models/kubernetes/chaosExperimentConnector'),
     config = require('../../../../src/common/consts').CONFIG;
 
 let manager;
@@ -49,11 +47,6 @@ describe('Manager jobs', function () {
 
     let testsManagerGetStub;
 
-    let getFutureJobExperimentsStub;
-    let getChaosExperimentByIdStub;
-
-    let runChaosExperimentStub;
-
     before(() => {
         sandbox = sinon.sandbox.create();
 
@@ -69,11 +62,6 @@ describe('Manager jobs', function () {
         failReportStub = sandbox.stub(reportsManager, 'failReport');
 
         testsManagerGetStub = sandbox.stub(testsManager, 'getTest');
-
-        getFutureJobExperimentsStub = sandbox.stub(chaosExperimentsManager, 'getFutureJobExperiments');
-        getChaosExperimentByIdStub = sandbox.stub(chaosExperimentsManager, 'getChaosExperimentById');
-
-        runChaosExperimentStub = sandbox.stub(chaosExperimentConnector, 'runChaosExperiment');
 
         jobGetLogsStub = sandbox.stub(jobConnector, 'getLogs');
         jobDeleteContainerStub = sandbox.stub(jobConnector, 'deleteAllContainers');
@@ -170,33 +158,6 @@ describe('Manager jobs', function () {
                     }
                 }, 3000);
             });
-        });
-    });
-
-    describe('Reload job experiments', function () {
-        it('found future experiments to reload', async () => {
-            const timestamp = 500;
-            const jobExperiment = { start_time: timestamp, job_id: '1234', experiment_id: '4321', id: '2468' };
-            const chaosExperiment = { kubeObject: { hello: 1 }, experiment_id: '4321' };
-            getFutureJobExperimentsStub.resolves([jobExperiment]);
-            getChaosExperimentByIdStub.resolves(chaosExperiment);
-            runChaosExperimentStub.returns();
-
-            const clock = sinon.useFakeTimers();
-            clock.tick(1000);
-            await manager.reloadChaosExperiments();
-            clock.tick(3000);
-            sinon.assert.calledOnce(runChaosExperimentStub);
-            sinon.assert.calledWith(runChaosExperimentStub, chaosExperiment.kubeObject);
-            clock.restore();
-        });
-        it('future experiments not found - nothing to reload', async () => {
-            getFutureJobExperimentsStub.resolves([]);
-            runChaosExperimentStub.returns();
-
-            await manager.reloadChaosExperiments();
-            sinon.assert.notCalled(getChaosExperimentByIdStub);
-            sinon.assert.notCalled(runChaosExperimentStub);
         });
     });
