+++ tests/unit-tests/k8sChaosExperiments/models/chaosExperimentsManager-test.js
@@ -11,7 +11,6 @@ const chaosExperimentConnector = require('../../../../src/chaos-experiments/mode
 describe('Chaos experiments manager tests', function () {
     let sandbox;
     let deleteStub;
-    let getChaosExperimentByIdStub;
     let getChaosExperimentByNameStub;
     let getChaosExperimentsStub;
     let getChaosExperimentsByIdsStub;
@@ -19,6 +18,8 @@ describe('Chaos experiments manager tests', function () {
     let insertStub;
     let setChaosJobExperimentTriggeredStub;
     let runChaosExperimentConnectorStub;
+    let getFutureJobExperimentsStub;
+    let getChaosExperimentByIdStub;
 
     before(() => {
         sandbox = sinon.sandbox.create();
@@ -31,6 +32,7 @@ describe('Chaos experiments manager tests', function () {
         updatedChaosExperimentStub = sandbox.stub(database, 'updateChaosExperiment');
         setChaosJobExperimentTriggeredStub = sandbox.stub(database, 'setChaosJobExperimentTriggered');
         runChaosExperimentConnectorStub = sandbox.stub(chaosExperimentConnector, 'runChaosExperiment');
+        getFutureJobExperimentsStub = sandbox.stub(database, 'getFutureJobExperiments');
     });
     after(() => {
         sandbox.restore();
@@ -377,6 +379,33 @@ describe('Chaos experiments manager tests', function () {
             runChaosExperimentConnectorStub.args[0][0].should.eql(kubernetesJobConfig);
             setChaosJobExperimentTriggeredStub.calledOnce.should.eql(true);
             setChaosJobExperimentTriggeredStub.args[0][0].should.eql(chaosJobExperimentId);
+        });
+    });
+
+    describe('Reload job experiments', function () {
+        it('found future experiments to reload', async () => {
+            const timestamp = 500;
+            const jobExperiment = { start_time: timestamp, job_id: '1234', experiment_id: '4321', id: '2468' };
+            const chaosExperiment = { kubeObject: { hello: 1 }, experiment_id: '4321' };
+            getFutureJobExperimentsStub.resolves([jobExperiment]);
+            getChaosExperimentByIdStub.resolves(chaosExperiment);
+            runChaosExperimentConnectorStub.returns();
+
+            const clock = sinon.useFakeTimers();
+            clock.tick(1000);
+            await manager.reloadChaosExperiments();
+            clock.tick(3000);
+            sinon.assert.calledOnce(runChaosExperimentConnectorStub);
+            sinon.assert.calledWith(runChaosExperimentConnectorStub, chaosExperiment.kubeObject);
+            clock.restore();
+        });
+        it('future experiments not found - nothing to reload', async () => {
+            getFutureJobExperimentsStub.resolves([]);
+            runChaosExperimentConnectorStub.returns();
+
+            await manager.reloadChaosExperiments();
+            sinon.assert.notCalled(getChaosExperimentByIdStub);
+            sinon.assert.notCalled(runChaosExperimentConnectorStub);
         });
     });
 });
